Move all tests to the same file
