+++ app/tests/algorithms_tests/test_api.py
@@ -13,8 +13,6 @@ from grandchallenge.algorithms.models import Endpoint, Job
 from grandchallenge.algorithms.serializers import (
     AlgorithmImageSerializer,
     AlgorithmModelSerializer,
-    HyperlinkedInvocationSerializer,
-    InvocationPostSerializer,
 )
 from grandchallenge.components.models import (
     ComponentInterfaceValue,
@@ -27,7 +25,6 @@ from tests.algorithms_tests.factories import (
     AlgorithmJobFactory,
     AlgorithmModelFactory,
     EndpointFactory,
-    InvocationFactory,
 )
 from tests.cases_tests import RESOURCE_PATH
 from tests.cases_tests.factories import (
@@ -809,44 +806,3 @@ class TestEndpointReadOnly:
         client.force_login(user=endpoint.creator)
         response = client.delete(f"{self.url}{endpoint.pk}/")
         assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_hyperlinked_invocation_serializer(rf):
-    invocation = InvocationFactory()
-    civ = ComponentInterfaceValueFactory(
-        interface__kind=InterfaceKindChoices.PANIMG_IMAGE, image=ImageFactory()
-    )
-    invocation.inputs.set([civ])
-    serializer = HyperlinkedInvocationSerializer(
-        invocation, context={"request": rf.get("/foo", secure=True)}
-    )
-    assert serializer.data["status"] == invocation.get_status_display()
-    assert serializer.data["endpoint"] == invocation.endpoint.api_url
-    assert serializer.data["inputs"][0]["image"] == str(
-        invocation.inputs.first().image.api_url
-    )
-
-
-@pytest.mark.django_db
-def test_invocation_post_serializer_endpoint_queryset(rf):
-    user = UserFactory()
-    endpoint_with_perm_running = EndpointFactory(
-        creator=user, status=Endpoint.StatusChoices.RUNNING
-    )
-    endpoint_with_perm_pending = EndpointFactory(creator=user)
-    endpoint_without_perm = EndpointFactory(
-        status=Endpoint.StatusChoices.RUNNING
-    )
-    request = rf.get("/foo", secure=True)
-    request.user = user
-
-    serializer = InvocationPostSerializer(
-        InvocationFactory(), context={"request": request}
-    )
-    assert endpoint_with_perm_running in serializer.fields["endpoint"].queryset
-    assert (
-        endpoint_with_perm_pending
-        not in serializer.fields["endpoint"].queryset
-    )
-    assert endpoint_without_perm not in serializer.fields["endpoint"].queryset
